Introduce MCMC, NLP Lab
Introduction of Markov Chain Monty Carlo with code along. Will need to update with the appropriate fixes in order to make sure the code is properly commented and runs well.

NLP lab consisted of reviewing whether or not text present contained an insult or not. This was accomplished through nltk analysis of word count, passing through the classifier. Used pipelines to perform the analysis.