Complete Indeed Data Analysis
Analyzed the indeed data of job postings for data related fields. Used tfidf to vectorize the words, then analyzed subsets for data science specifically and not specifically.

Key point being creation of a dot plot with several variables to indicate the relative frequency of the ngrams in the job postings.